+++ mne/viz/tests/test_3d.py
@@ -21,7 +21,7 @@ from mne.io.meas_info import write_dig
 from mne.viz import (plot_sparse_source_estimates, plot_source_estimates,
                      plot_trans, snapshot_brain_montage)
 from mne.utils import (requires_mayavi, requires_pysurfer, run_tests_if_main,
-                       _import_mlab, _TempDir)
+                       _import_mlab, _TempDir, requires_nibabel)
 from mne.datasets import testing
 from mne.source_space import read_source_spaces
 
@@ -256,8 +256,32 @@ def test_plot_dipole_locations():
                   subjects_dir, mode='foo')
 
 
+@testing.requires_testing_data
+@requires_nibabel
+def test_plot_dipole_mri_orthoview():
+    """Test mpl dipole plotting."""
+    import matplotlib.pyplot as plt
+    dipoles = read_dipole(dip_fname)
+    trans = read_trans(trans_fname)
+    for coord_frame, idx, show_all in zip(['head', 'mri'],
+                                          ['gof', 'amplitude'], [True, False]):
+        fig = dipoles.plot_locations(trans, 'sample', subjects_dir,
+                                     coord_frame=coord_frame, idx=idx,
+                                     show_all=show_all, mode='orthoview')
+        fig.canvas.scroll_event(0.5, 0.5, 1)  # scroll up
+        fig.canvas.scroll_event(0.5, 0.5, -1)  # scroll down
+        fig.canvas.key_press_event('up')
+        fig.canvas.key_press_event('down')
+        fig.canvas.key_press_event('a')  # some other key
+    ax = plt.subplot(111)
+    assert_raises(ValueError, dipoles.plot_locations, trans, 'sample',
+                  subjects_dir, ax=ax)
+    plt.close('all')
+
+
 @requires_mayavi
 def test_snapshot_brain_montage():
+    """Test snapshot brain montage."""
     info = read_info(evoked_fname)
     fig = plot_trans(info, trans=None, subject='sample',
                      subjects_dir=subjects_dir)
@@ -275,5 +299,6 @@ def test_snapshot_brain_montage():
 
     # Make sure we raise error if the figure has no scene
     assert_raises(TypeError, snapshot_brain_montage, fig, info)
+
 
 run_tests_if_main()